simplify Ehcache Java Configuration
