+++ src/main/resources/generator/server/springboot/cache/ehcache/src/CacheConfiguration.java.mustache
@@ -1,0 +1,62 @@
+package {{packageName}}.technical.infrastructure.secondary.cache;
+
+import java.time.Duration;
+import javax.cache.Cache;
+import javax.cache.CacheManager;
+import org.ehcache.config.builders.CacheConfigurationBuilder;
+import org.ehcache.config.builders.ExpiryPolicyBuilder;
+import org.ehcache.config.builders.ResourcePoolsBuilder;
+import org.ehcache.jsr107.Eh107Configuration;
+import org.springframework.boot.autoconfigure.cache.JCacheManagerCustomizer;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableCaching
+public class CacheConfiguration {
+
+  private final javax.cache.configuration.Configuration<Object, Object> configuration;
+
+  public CacheConfiguration(EhcacheProperties properties) {
+    this.configuration =
+      Eh107Configuration.fromEhcacheCacheConfiguration(
+        CacheConfigurationBuilder
+          .newCacheConfigurationBuilder(Object.class, Object.class, ResourcePoolsBuilder.heap(properties.getMaxEntries()))
+          .withExpiry(ExpiryPolicyBuilder.timeToLiveExpiration(Duration.ofSeconds(properties.getTimeToLiveSeconds())))
+          .build()
+      );
+  }
+
+  @Bean
+  public JCacheManagerCustomizer cacheManagerCustomizer() {
+    return cm -> {
+      // jhipster-needle-jcache-add-entity
+    };
+  }
+
+  /**
+   * Create or replace a cache with the default configuration
+   *
+   * @param cm        cache manager
+   * @param cacheName cache name
+   */
+  void createCache(CacheManager cm, String cacheName) {
+    createCache(cm, cacheName, configuration);
+  }
+
+  /**
+   * Create or replace a cache
+   *
+   * @param cm            cache manager
+   * @param cacheName     cache name
+   * @param configuration cache configuration
+   */
+  static void createCache(CacheManager cm, String cacheName, javax.cache.configuration.Configuration<Object, Object> configuration) {
+    Cache<Object, Object> cache = cm.getCache(cacheName);
+    if (cache != null) {
+      cm.destroyCache(cacheName);
+    }
+    cm.createCache(cacheName, configuration);
+  }
+}
