--- src/main/resources/generator/server/springboot/cache/ehcache/src/EhcacheConfiguration.java.mustache
@@ -1,19 +1,0 @@
-package {{packageName}}.technical.infrastructure.secondary.cache.ehcache;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class EhcacheConfiguration {
-
-  private final EhcacheProperties properties;
-
-  public EhcacheConfiguration(EhcacheProperties properties) {
-    this.properties = properties;
-  }
-
-  @Bean
-  public EhcacheConfigurer ehcacheConfigurer() {
-    return new EhcacheConfigurer(properties);
-  }
-}
