+++ src/main/resources/generator/server/springboot/cache/ehcache/src/EhcacheProperties.java.mustache
@@ -1,4 +1,4 @@
-package {{packageName}}.technical.infrastructure.secondary.cache.ehcache;
+package {{packageName}}.technical.infrastructure.secondary.cache;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
