+++ src/main/resources/generator/server/springboot/cache/ehcache/test/CacheConfigurationTest.java.mustache
@@ -1,0 +1,53 @@
+package {{packageName}}.technical.infrastructure.secondary.cache;
+
+import static org.mockito.Mockito.*;
+
+import javax.cache.Cache;
+import javax.cache.CacheManager;
+import javax.cache.configuration.Configuration;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import {{packageName}}.UnitTest;
+
+@UnitTest
+@ExtendWith(MockitoExtension.class)
+class CacheConfigurationTest {
+
+  private static final CacheConfiguration cacheConfiguration = new CacheConfiguration(new EhcacheProperties());
+
+  @Mock
+  CacheManager cm;
+
+  @Mock
+  Cache<Object, Object> cache;
+
+  @Mock
+  Configuration<Object, Object> customConfig;
+
+  @Test
+  void shouldCreateDefaultCache() {
+    cacheConfiguration.createCache(cm, "shouldCreateDefaultCache");
+    verify(cm).createCache(eq("shouldCreateDefaultCache"), any());
+  }
+
+  @Test
+  void shouldCreateCustomCache() {
+    CacheConfiguration.createCache(cm, "shouldCreateCustomCache", customConfig);
+    verify(cm).createCache("shouldCreateCustomCache", customConfig);
+  }
+
+  @Test
+  void shouldReplaceExistingCache() {
+    // given an existing cache "test"
+    doReturn(cache).when(cm).getCache("shouldReplaceExistingCache");
+
+    // when attempting to create a new cache "test"
+    CacheConfiguration.createCache(cm, "shouldReplaceExistingCache", customConfig);
+
+    // then destroy and create cache "test"
+    verify(cm).destroyCache("shouldReplaceExistingCache");
+    verify(cm).createCache("shouldReplaceExistingCache", customConfig);
+  }
+}
