--- src/main/resources/generator/server/springboot/cache/jcache/src/JCacheConfiguration.java.mustache
@@ -1,23 +1,0 @@
-package {{packageName}}.technical.infrastructure.secondary.cache;
-
-import org.springframework.boot.autoconfigure.cache.JCacheManagerCustomizer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JCacheConfiguration {
-
-  @SuppressWarnings("java:S1068") // Not used without entity
-  private final JCacheCreator jCacheCreator;
-
-  public JCacheConfiguration(JCacheCreator jCacheCreator) {
-    this.jCacheCreator = jCacheCreator;
-  }
-
-  @Bean
-  public JCacheManagerCustomizer cacheManagerCustomizer() {
-    return cm -> {
-      // jhipster-needle-jcache-add-entity
-    };
-  }
-}
