--- src/main/resources/generator/server/springboot/cache/jcache/src/JCacheConfigurer.java.mustache
@@ -1,8 +1,0 @@
-package {{packageName}}.technical.infrastructure.secondary.cache;
-
-import javax.cache.configuration.Configuration;
-
-@FunctionalInterface
-public interface JCacheConfigurer {
-  Configuration<Object, Object> configure();
-}
