--- src/test/java/tech/jhipster/lite/generator/server/springboot/cache/jcache/application/SpringBootJCacheApplicationServiceIT.java
@@ -1,71 +1,0 @@
-package tech.jhipster.lite.generator.server.springboot.cache.jcache.application;
-
-import static tech.jhipster.lite.TestUtils.tmpProject;
-import static tech.jhipster.lite.generator.project.domain.DefaultConfig.BASE_NAME;
-import static tech.jhipster.lite.generator.project.domain.DefaultConfig.PACKAGE_NAME;
-import static tech.jhipster.lite.generator.server.springboot.cache.jcache.application.SpringBootJCacheAssert.*;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import tech.jhipster.lite.IntegrationTest;
-import tech.jhipster.lite.generator.buildtool.maven.application.MavenApplicationService;
-import tech.jhipster.lite.generator.init.application.InitApplicationService;
-import tech.jhipster.lite.generator.project.domain.Project;
-import tech.jhipster.lite.generator.server.springboot.core.application.SpringBootApplicationService;
-
-@IntegrationTest
-class SpringBootJCacheApplicationServiceIT {
-
-  @Autowired
-  SpringBootJCacheApplicationService springBootJCacheApplicationService;
-
-  @Autowired
-  InitApplicationService initApplicationService;
-
-  @Autowired
-  MavenApplicationService mavenApplicationService;
-
-  @Autowired
-  SpringBootApplicationService springBootApplicationService;
-
-  @Test
-  void shouldAddDependencies() {
-    Project project = tmpProject();
-    project.addConfig(BASE_NAME, "bar");
-    project.addConfig(PACKAGE_NAME, "tech.jhipster.baz");
-    initApplicationService.init(project);
-    mavenApplicationService.addPomXml(project);
-
-    springBootJCacheApplicationService.addDependencies(project);
-
-    assertDependencies(project);
-  }
-
-  @Test
-  void shouldEnableCaching() {
-    Project project = tmpProject();
-    project.addConfig(BASE_NAME, "bar");
-    project.addConfig(PACKAGE_NAME, "tech.jhipster.baz");
-    initApplicationService.init(project);
-    mavenApplicationService.addPomXml(project);
-    springBootApplicationService.init(project);
-
-    springBootJCacheApplicationService.addEnableCaching(project);
-
-    assertEnableCaching(project);
-  }
-
-  @Test
-  void shouldAddJavaConfig() {
-    Project project = tmpProject();
-    project.addConfig(BASE_NAME, "bar");
-    project.addConfig(PACKAGE_NAME, "tech.jhipster.baz");
-    initApplicationService.init(project);
-    mavenApplicationService.addPomXml(project);
-    springBootApplicationService.init(project);
-
-    springBootJCacheApplicationService.addJavaConfig(project);
-
-    assertJavaConfig(project);
-  }
-}
